Revert the previous ULONG -> SIZE_T changes and cast affected variables to ULONG instead.
Changes to inflib affect many dependencies, some of which are unchangeable Win32 APIs (e.g. setupapi.dll).
Keeping 32-bit sizes everywhere is better than pretending we support 64-bit SIZE_T and then casting down to 32-bit anyway.

+++ sdk/lib/inflib/infhostput.c
@@ -18,7 +18,7 @@ InfHostWriteFile(HINF InfHandle,
                  const CHAR *HeaderComment)
 {
   WCHAR *Buffer;
-  SIZE_T BufferSize;
+  ULONG BufferSize;
   INFSTATUS Status;
   FILE *File;
